refactor: remove SlidingPillShell component and integrate its functionality into Tabs for improved navigation
